refactor: DUMMY 수정 및 data 패키지로 위치 변경
